father/requirements.txt
bs4 
google-cloud-pubsub
google-cloud-storage
requests
google-cloud-error_reporting
unidecode